refactor(handle): naming things, cleanup
